paintress-server/pyproject.toml
[project]
name = "paintress-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "aiofiles>=24.1.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "clerk-backend-api>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]
